move test models to UUID, add default id key, cleanup test structure
