+++ Sources/FluentBenchmark/FluentBenchmarker.swift
@@ -4,6 +4,8 @@ import NIO
 import XCTest
 
 public final class FluentBenchmarker {
+    public static var idKey: String = "id"
+
     public let database: Database
     
     public init(database: Database) {
