+++ Sources/FluentBenchmark/SolarSystem/Galaxy.swift
@@ -3,7 +3,7 @@ import FluentKit
 public final class Galaxy: Model {
     public static let schema = "galaxies"
     
-    @ID(key: "id")
+    @ID(key: FluentBenchmarker.idKey)
     public var id: UUID?
 
     @Field(key: "name")
