+++ Sources/FluentBenchmark/Tests/Enum.swift
@@ -1,39 +1,7 @@
 extension FluentBenchmarker {
     public func testUInt8BackedEnum() throws {
-        enum Bar: UInt8, Codable {
-            case baz, qux
-        }
-        final class Foo: Model {
-            static let schema = "foos"
-
-            struct _Migration: Migration {
-                func prepare(on database: Database) -> EventLoopFuture<Void> {
-                    return database.schema("foos")
-                        .field("id", .int, .identifier(auto: true))
-                        .field("bar", .uint8, .required)
-                        .create()
-                }
-
-                func revert(on database: Database) -> EventLoopFuture<Void> {
-                    return database.schema("foos").delete()
-                }
-            }
-
-            @ID(key: "id")
-            var id: Int?
-
-            @Field(key: "bar")
-            var bar: Bar
-
-            init() { }
-
-            init(id: Int? = nil, bar: Bar) {
-                self.id = id
-                self.bar = bar
-            }
-        }
         try runTest(#function, [
-            Foo._Migration()
+            FooMigration()
         ]) {
             let foo = Foo(bar: .baz)
             try foo.save(on: self.database).wait()
@@ -41,5 +9,40 @@ extension FluentBenchmarker {
             let fetched = try Foo.find(foo.id, on: self.database).wait()
             XCTAssertEqual(fetched?.bar, .baz)
         }
+    }
+}
+
+private enum Bar: UInt8, Codable {
+    case baz, qux
+}
+
+private final class Foo: Model {
+    static let schema = "foos"
+
+    @ID(key: FluentBenchmarker.idKey)
+    var id: UUID?
+
+    @Field(key: "bar")
+    var bar: Bar
+
+    init() { }
+
+    init(id: IDValue? = nil, bar: Bar) {
+        self.id = id
+        self.bar = bar
+    }
+}
+
+
+private struct FooMigration: Migration {
+    func prepare(on database: Database) -> EventLoopFuture<Void> {
+        database.schema("foos")
+            .field("id", .uuid, .identifier(auto: false))
+            .field("bar", .uint8, .required)
+            .create()
+    }
+
+    func revert(on database: Database) -> EventLoopFuture<Void> {
+        database.schema("foos").delete()
     }
 }
