+++ Sources/FluentBenchmark/Tests/Migrator.swift
@@ -49,3 +49,17 @@ extension FluentBenchmarker {
         }
     }
 }
+
+private struct ErrorMigration: Migration {
+    init() { }
+
+    struct Error: Swift.Error { }
+
+    func prepare(on database: Database) -> EventLoopFuture<Void> {
+        return database.eventLoop.makeFailedFuture(Error())
+    }
+
+    func revert(on database: Database) -> EventLoopFuture<Void> {
+        return database.eventLoop.makeSucceededFuture(())
+    }
+}
