+++ Sources/FluentBenchmark/Tests/NestedModel.swift
@@ -26,7 +26,7 @@ extension FluentBenchmarker {
             }
 
             struct UserJSON: Equatable, Codable {
-                var id: Int
+                var id: UUID
                 var name: String
                 var pet: PetJSON
             }
@@ -34,8 +34,8 @@ extension FluentBenchmarker {
                 var name: String
                 var type: String
             }
-            // {"id":2,"name":"Tanner","pet":{"name":"Ziz","type":"cat"}}
-            let expected = UserJSON(id: 2, name: "Tanner", pet: .init(name: "Ziz", type: "cat"))
+            // {"id":...,"name":"Tanner","pet":{"name":"Ziz","type":"cat"}}
+            let expected = UserJSON(id: user.id!, name: "Tanner", pet: .init(name: "Ziz", type: "cat"))
 
             let decoded = try JSONDecoder().decode(UserJSON.self, from: JSONEncoder().encode(user))
             guard decoded == expected else {
@@ -43,5 +43,72 @@ extension FluentBenchmarker {
                 return
             }
         }
+    }
+}
+
+private final class User: Model {
+    struct Pet: Codable {
+        enum Animal: String, Codable {
+            case cat, dog
+        }
+        var name: String
+        var type: Animal
+    }
+    static let schema = "users"
+
+    @ID(key: FluentBenchmarker.idKey)
+    var id: UUID?
+
+    @Field(key: "name")
+    var name: String
+
+    @Field(key: "pet")
+    var pet: Pet
+
+    @OptionalParent(key: "bf_id")
+    var bestFriend: User?
+
+    @Children(for: \.$bestFriend)
+    var friends: [User]
+
+    init() { }
+
+    init(id: IDValue? = nil, name: String, pet: Pet, bestFriend: User? = nil) {
+        self.id = id
+        self.name = name
+        self.pet = pet
+        self.$bestFriend.id = bestFriend?.id
+    }
+}
+
+private struct UserMigration: Migration {
+    func prepare(on database: Database) -> EventLoopFuture<Void> {
+        database.schema("users")
+            .field("id", .uuid, .identifier(auto: false))
+            .field("name", .string, .required)
+            .field("pet", .json, .required)
+            .field("bf_id", .int)
+            .create()
+    }
+
+    func revert(on database: Database) -> EventLoopFuture<Void> {
+        database.schema("users").delete()
+    }
+}
+
+
+private struct UserSeed: Migration {
+    init() { }
+
+    func prepare(on database: Database) -> EventLoopFuture<Void> {
+        let tanner = User(name: "Tanner", pet: .init(name: "Ziz", type: .cat))
+        let logan = User(name: "Logan", pet: .init(name: "Runa", type: .dog))
+        return logan.save(on: database)
+            .and(tanner.save(on: database))
+            .map { _ in }
+    }
+
+    func revert(on database: Database) -> EventLoopFuture<Void> {
+        database.eventLoop.makeSucceededFuture(())
     }
 }
