+++ Sources/FluentBenchmark/Tests/Timestamp.swift
@@ -1,45 +1,5 @@
 extension FluentBenchmarker {
     public func testTimestampable() throws {
-        final class User: Model {
-            static let schema = "users"
-
-            @ID(key: "id")
-            var id: Int?
-
-            @Field(key: "name")
-            var name: String
-
-            @Timestamp(key: "created_at", on: .create)
-            var createdAt: Date?
-
-            @Timestamp(key: "updated_at", on: .update)
-            var updatedAt: Date?
-
-            init() { }
-            init(id: Int? = nil, name: String) {
-                self.id = id
-                self.name = name
-                self.createdAt = nil
-                self.updatedAt = nil
-            }
-        }
-
-        struct UserMigration: Migration {
-            func prepare(on database: Database) -> EventLoopFuture<Void> {
-                return database.schema("users")
-                    .field("id", .int, .identifier(auto: true))
-                    .field("name", .string, .required)
-                    .field("created_at", .datetime)
-                    .field("updated_at", .datetime)
-                    .create()
-            }
-
-            func revert(on database: Database) -> EventLoopFuture<Void> {
-                return database.schema("users").delete()
-            }
-        }
-
-
         try runTest(#function, [
             UserMigration(),
         ]) {
@@ -58,3 +18,44 @@ extension FluentBenchmarker {
         }
     }
 }
+
+private final class User: Model {
+    static let schema = "users"
+
+    @ID(key: FluentBenchmarker.idKey)
+    var id: UUID?
+
+    @Field(key: "name")
+    var name: String
+
+    @Timestamp(key: "created_at", on: .create)
+    var createdAt: Date?
+
+    @Timestamp(key: "updated_at", on: .update)
+    var updatedAt: Date?
+
+    init() { }
+    
+    init(id: IDValue? = nil, name: String) {
+        self.id = id
+        self.name = name
+        self.createdAt = nil
+        self.updatedAt = nil
+    }
+}
+
+private struct UserMigration: Migration {
+    func prepare(on database: Database) -> EventLoopFuture<Void> {
+        database.schema("users")
+            .field("id", .uuid, .identifier(auto: false))
+            .field("name", .string, .required)
+            .field("created_at", .datetime)
+            .field("updated_at", .datetime)
+            .create()
+    }
+
+    func revert(on database: Database) -> EventLoopFuture<Void> {
+        database.schema("users").delete()
+    }
+}
+
